Separate test utils for metrics into its own directory (#791)
Summary:
Pull Request resolved: https://github.com/pytorch/torchrec/pull/791

fixing module not found error

Reviewed By: colin2328

Differential Revision: D41091570

fbshipit-source-id: f005b09373ae1c052c8a17626042a1b605154d98
