+++ torchrec/metrics/tests/test_ne.py
@@ -14,7 +14,7 @@ import torch
 import torch.distributed as dist
 from torchrec.metrics.ne import compute_cross_entropy, compute_ne, NEMetric
 from torchrec.metrics.rec_metric import RecComputeMode, RecMetric
-from torchrec.metrics.tests.test_utils import (
+from torchrec.metrics.test_utils import (
     rec_metric_value_test_helper,
     rec_metric_value_test_launcher,
     TestMetric,
